update_20190621
update_20190621

+++ t5-neith/SipTool/source/neighborcfglogic.cpp
@@ -1,6 +1,7 @@
 #include "stdafx.h"
 #include "neighborcfglogic.h"
 #include "mainframelogic.h"
+#include "messageboxlogic.h"
 
 template<> CNeighborCfgLogic* Singleton<CNeighborCfgLogic>::ms_pSingleton  = NULL;
 
@@ -31,15 +32,35 @@ bool CNeighborCfgLogic::NeiRegServerItemAdd(TNeiRegServerInfo& tNeiRegServerInfo
 
     CString str;
     CListTextElementUI* pListElement = new CListTextElementUI;
-    pListElement->SetTag(nCount);
-    m_pNeiRegServerList->Add(pListElement);
+    if (pListElement)
+    {
+        pListElement->SetAttribute(_T("height"),_T("24"));
+        pListElement->SetAttribute(_T("gdiplus"),_T("true"));
+        pListElement->SetTag(nCount);
+        m_pNeiRegServerList->Add(pListElement);
 
-    str.Format(_T("%d"), nCount);
-    pListElement->SetText(0, str);
-    pListElement->SetText(1, (CA2T)tNeiRegServerInfo.m_achAreaCode);
-    pListElement->SetText(2, (CA2T)tNeiRegServerInfo.m_achIpAddr);
-    str.Format(_T("%d"), tNeiRegServerInfo.m_wPort);
-    pListElement->SetText(3, str);
+        str.Format(_T("%d"), nCount+1);
+        pListElement->SetText(0, str);
+        pListElement->SetText(1, (CA2T)tNeiRegServerInfo.m_achAreaCode);
+        pListElement->SetText(2, (CA2T)tNeiRegServerInfo.m_achIpAddr);
+        str.Format(_T("%d"), tNeiRegServerInfo.m_wPort);
+        pListElement->SetText(3, str);
+    }
+
+    return true;
+}
+
+bool CNeighborCfgLogic::AreaCodeIsExist(TNeiRegServerInfo& tNeiRegServerInfo)
+{
+    s32 nCount = m_pNeiRegServerList->GetCount();
+    for (s32 nIndex = 0; nIndex < nCount; nIndex++)
+    {
+        CListTextElementUI* pListElement = (CListTextElementUI*)m_pNeiRegServerList->GetItemAt(nIndex);
+        if ( strcmp( (CT2A)pListElement->GetText(1), tNeiRegServerInfo.m_achAreaCode ) == 0 )// 区号相同
+        {
+            return false;
+        }
+    }
 
     return true;
 }
@@ -54,10 +75,13 @@ bool CNeighborCfgLogic::OnNeighborCfgAddBtnClicked(TNotifyUI& msg)
 
 bool CNeighborCfgLogic::OnNeighborCfgDelBtnClicked(TNotifyUI& msg)
 {
-    s32 nCurSel = m_pNeiRegServerList->GetCurSel();
-    CListTextElementUI* pCurListElement = (CListTextElementUI*)m_pNeiRegServerList->GetItemAt(nCurSel);
-    CString cstrAreaNum(pCurListElement->GetText(1));
-    CSipToolComInterface->DeleteNeighborInfo( (CT2A)cstrAreaNum );
+    if ( ShowMessageBox(_T("是否删除？"), 2) == true )
+    {
+        s32 nCurSel = m_pNeiRegServerList->GetCurSel();
+        CListTextElementUI* pCurListElement = (CListTextElementUI*)m_pNeiRegServerList->GetItemAt(nCurSel);
+        CString cstrAreaNum(pCurListElement->GetText(1));
+        CSipToolComInterface->DeleteNeighborInfo( (CT2A)cstrAreaNum );
+    }
 
     return true;
 }
@@ -78,15 +102,20 @@ bool CNeighborCfgLogic::OnUpdateNeighborInfoList( WPARAM wparam, LPARAM lparam, 
         for (u32 dwIndex = 0; dwIndex < dwNeighborInfoSize; dwIndex++)
         {
             CListTextElementUI* pListElement = new CListTextElementUI;
-            pListElement->SetTag(dwIndex);
-            m_pNeiRegServerList->Add(pListElement);
+            if (pListElement)
+            {
+                pListElement->SetAttribute(_T("height"),_T("24"));
+                pListElement->SetAttribute(_T("gdiplus"),_T("true"));
+                pListElement->SetTag(dwIndex);
+                m_pNeiRegServerList->Add(pListElement);
 
-            str.Format(_T("%d"), dwIndex);
-            pListElement->SetText(0, str);
-            pListElement->SetText(1, (CA2T)vNeighborInfo[dwIndex].m_achAreaCode);
-            pListElement->SetText(2, (CA2T)vNeighborInfo[dwIndex].m_achIpAddr);
-            str.Format(_T("%d"), vNeighborInfo[dwIndex].m_wPort);
-            pListElement->SetText(3, str);
+                str.Format(_T("%d"), dwIndex+1);
+                pListElement->SetText(0, str);
+                pListElement->SetText(1, (CA2T)vNeighborInfo[dwIndex].m_achAreaCode);
+                pListElement->SetText(2, (CA2T)vNeighborInfo[dwIndex].m_achIpAddr);
+                str.Format(_T("%d"), vNeighborInfo[dwIndex].m_wPort);
+                pListElement->SetText(3, str);
+            }
 
             //::delete pListElement;
             //pListElement = NULL;
@@ -109,9 +138,11 @@ bool CNeighborCfgLogic::OnDeleteNeighborInfoList( WPARAM wparam, LPARAM lparam, 
         for (s32 nIndex = nCurSel; nIndex < m_pNeiRegServerList->GetCount(); nIndex++)
         {
             CListTextElementUI* pListElement = (CListTextElementUI*)m_pNeiRegServerList->GetItemAt(nIndex);
-            strSerial.Format(_T("%d"), nIndex);
+            strSerial.Format(_T("%d"), nIndex+1);
             pListElement->SetText(0, strSerial);
         }
+
+        SHOWTIP(_T("邻居信息删除成功！"));
     }
 
     return true;
